+++ noodles-bam/src/reader/record/data/field/value/subtype.rs
@@ -1,16 +1,17 @@
-use std::io::{self, Read};
+use std::{io, mem};
 
-use byteorder::ReadBytesExt;
-
+use bytes::Buf;
 use noodles_sam::record::data::field::value::Subtype;
 
-pub fn read_subtype<R>(reader: &mut R) -> io::Result<Subtype>
+pub fn get_subtype<B>(src: &mut B) -> io::Result<Subtype>
 where
-    R: Read,
+    B: Buf,
 {
-    reader.read_u8().and_then(|n| {
-        Subtype::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-    })
+    if src.remaining() < mem::size_of::<u8>() {
+        return Err(io::Error::from(io::ErrorKind::UnexpectedEof));
+    }
+
+    Subtype::try_from(src.get_u8()).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
 }
 
 #[cfg(test)]
@@ -18,15 +19,15 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_read_subtype() -> io::Result<()> {
+    fn test_get_subtype() -> io::Result<()> {
         let data = [b'i'];
         let mut reader = &data[..];
-        assert_eq!(read_subtype(&mut reader)?, Subtype::Int32);
+        assert_eq!(get_subtype(&mut reader)?, Subtype::Int32);
 
         let data = [b'n'];
         let mut reader = &data[..];
         assert!(matches!(
-            read_subtype(&mut reader),
+            get_subtype(&mut reader),
             Err(ref e) if e.kind() == io::ErrorKind::InvalidData
         ));