+++ noodles-bam/src/reader/record/data/field/value/ty.rs
@@ -1,16 +1,17 @@
-use std::io::{self, Read};
+use std::{io, mem};
 
-use byteorder::ReadBytesExt;
-
+use bytes::Buf;
 use noodles_sam::record::data::field::value::Type;
 
-pub fn read_type<R>(reader: &mut R) -> io::Result<Type>
+pub fn get_type<B>(src: &mut B) -> io::Result<Type>
 where
-    R: Read,
+    B: Buf,
 {
-    reader
-        .read_u8()
-        .and_then(|n| Type::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)))
+    if src.remaining() < mem::size_of::<u8>() {
+        return Err(io::Error::from(io::ErrorKind::UnexpectedEof));
+    }
+
+    Type::try_from(src.get_u8()).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
 }
 
 #[cfg(test)]
@@ -18,15 +19,15 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_read_type() -> io::Result<()> {
+    fn test_get_type() -> io::Result<()> {
         let data = [b'i'];
         let mut reader = &data[..];
-        assert_eq!(read_type(&mut reader)?, Type::Int32);
+        assert_eq!(get_type(&mut reader)?, Type::Int32);
 
         let data = [b'n'];
         let mut reader = &data[..];
         assert!(matches!(
-            read_type(&mut reader),
+            get_type(&mut reader),
             Err(ref e) if e.kind() == io::ErrorKind::InvalidData
         ));