+++ noodles-bam/src/record/data/fields.rs
@@ -1,34 +1,25 @@
-use std::io::{self, BufRead};
+use std::io;
 
-use crate::{reader::record::data::read_field, record::data::Field};
+use crate::{reader::record::data::get_field, record::data::Field};
 
 /// An iterator over BAM record data fields.
 ///
 /// This is created by calling [`crate::record::Data::fields`].
-pub struct Fields<R>
-where
-    R: BufRead,
-{
-    inner: R,
+pub struct Fields<'a> {
+    buf: &'a [u8],
 }
 
-impl<R> Fields<R>
-where
-    R: BufRead,
-{
-    pub(crate) fn new(inner: R) -> Self {
-        Self { inner }
+impl<'a> Fields<'a> {
+    pub(crate) fn new(buf: &'a [u8]) -> Self {
+        Self { buf }
     }
 }
 
-impl<R> Iterator for Fields<R>
-where
-    R: BufRead,
-{
+impl<'a> Iterator for Fields<'a> {
     type Item = io::Result<Field>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        match read_field(&mut self.inner) {
+        match get_field(&mut self.buf) {
             Ok(Some(field)) => Some(Ok(field)),
             Ok(None) => None,
             Err(e) => Some(Err(e)),