+++ noodles-cram/src/reader/record.rs
@@ -452,7 +452,7 @@ where
     }
 
     fn read_tag_data(&mut self) -> io::Result<Vec<Tag>> {
-        use bam::reader::record::data::field::read_value;
+        use bam::reader::record::data::field::get_value;
 
         let tag_line = self.read_tag_line().and_then(|i| {
             usize::try_from(i).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
@@ -486,7 +486,7 @@ where
             )?;
 
             let mut data_reader = &data[..];
-            let value = read_value(&mut data_reader, key.ty())?;
+            let value = get_value(&mut data_reader, key.ty())?;
 
             let tag = Tag::new(*key, value);
             tags.push(tag);